Refactor formatBytes util and use it on JobDetailPage
